pyproject.toml
[project]
name = "selfory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "dj-rest-auth[with-social]>=7.0.1",
    "django>=5.2.6",
    "django-countries>=7.6.1",
    "django-phonenumber-field>=8.1.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "environs==14.3.0",
    "phonenumbers>=9.0.14",
    "pillow>=11.3.0",
    "redis>=6.4.0",
]
